блокировки pmm и kheap

+++ kernel/arch/x86_64/memory/mem_layout.h
@@ -13,10 +13,10 @@
 
 #include "types.h"
 
-#define K2P(a) (void*)((uintptr_t)(a) & ~KERNEL_TEXT_OFFSET)
-#define P2K(a) (void*)((uintptr_t)(a) | KERNEL_TEXT_OFFSET)
+#define K2P(a) (void*)((uint64_t)(a) & ~KERNEL_TEXT_OFFSET)
+#define P2K(a) (void*)((uint64_t)(a) | KERNEL_TEXT_OFFSET)
 
-#define V2P(a) (void*)((uintptr_t)(a) & ~PHYSICAL_MEM_MAP_OFFSET)
-#define P2V(a) (void*)((uintptr_t)(a) | PHYSICAL_MEM_MAP_OFFSET)
+#define V2P(a) (void*)((uint64_t)(a) & ~PHYSICAL_MEM_MAP_OFFSET)
+#define P2V(a) (void*)((uint64_t)(a) | PHYSICAL_MEM_MAP_OFFSET)
 
 #endif
